Fix channel search selection issue, add predicate to channel list db observing
